Refactor code to use unique pointers in data processing
This commit refactors the code base to leverage the use of unique pointers within the data processing functions. Each function that processes a line, such as `processVertex`, `processFace`, and `processTexture`, now receives an additional parameter, which is a reference to a unique pointer to an `Item` object. In addition, `readAndPass` now utilizes `make_ptr` to create `Item` instances, and `common.h` was added to define this `make_ptr`.