Add scale factor and center point to Item class
The scale factor and center point are added to the 'Item' class for better 3D transformations. These new properties allow changes in the scale and location of the objects. The vertex reading is also modified to consider these new properties. Multiple 3D objects with different scale factor and center point can now be managed efficiently in the main() function.