Refactor ray scattering and camera functions
This commit includes significant refactoring of ray scattering functions in ray.h and ray.cpp to simplify intensity handling. Array properties used in ray scattering are no longer pointers, which simplifies the code and potentially improves performance. It also includes modification of the shootRaysOut method in the Camera class to return a pointer instead of a unique_ptr to simplify dereferencing. Minor changes have been applied to other parts of the code to accommodate these changes.