Add CoordTransform class and improve camera functionality
Added the CoordTransform class, which includes methods for setting points and calculating rotation matrix. This class was also included in the camera header file, implying its use in the camera functionality. Additionally, the 'overlappingFactor' function in camera.cpp was changed to be an inline function. Modifications were also made in CMakeLists.txt to include the new files and changes.