Improve Ray tracing logic and optimize mathematical conversions
Refined the ray tracing approach in camera.cpp by adding the `shootRayRandom` function and modifying the if statement in `shootRay` function. Substituted constants `M_PI` with `std::numbers::pi` in trans_forms.cpp for better readability and precision. The 'intersectsWithBox' function in 'ray.h' is now marked as [[nodiscard]] to prevent misuses. Adjustments are also made in main.cpp to utilize the new ray tracing approach.