Make classes final and add license in header files
The commit introduces the keyword 'final' to all classes in the project to prevent further inheritance. It also adds a copyright notice that includes a reference to the MIT license in the header files 'vec.h' and 'point.h'. Minor adjustments were made to some header inclusions in 'common.h' for better organization.