Update compatibility for C++17 and adjust multithreading
This update introduces backward compatibility for the C++17 standard and adapts code snippets for proper execution under this standard. Some conditional compiling directives have been added to facilitate this compatibility. Also, adjustments have been made to the multithreading process. This includes the introduction of an atomic datatype and the modification of the logging system to accommodate multithreaded operations. Furthermore, insignificant headers and macros were removed for code cleanliness.