Add normal processing and item boundary update functionality
The commit introduces processing of normal vec from the obj file instead of computing it manually. It also adds functionality to update the physical boundary of the item based on face raw coordinates. Additional changes include minor tweak in the ray scattering mechanism, and some non-performance affecting clean-ups in the rendering logic.