Refactor code to segregate material data and improve ray scattering logic
The commit introduces a new class, MTLDataset, to encapsulate the material parameters that were previously embedded directly in the Item class. The scattering process of rays is also markedly enhanced by implementing the mtlDataset in the calculation of the scattered light intensity. Some other optimization tweaks have been applied to the existing code as well. The new design significantly improves the encapsulation and makes the ray scattering more physically accurate.